ui: update history length when we can

+++ src/libgpaste/ui/gpaste-ui-panel.h
@@ -35,7 +35,7 @@ G_PASTE_FINAL_TYPE (UiPanel, ui_panel, UI_PANEL, GtkBox)
 
 void g_paste_ui_panel_update_history_length (GPasteUiPanel *self,
                                              const gchar   *history,
-                                             guint32        length);
+                                             gsize          length);
 
 GtkWidget *g_paste_ui_panel_new (GPasteClient   *client,
                                  GPasteSettings *settings,
